# deposit-desk/catalog/providers/deposit-desk/rentals/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Deposit Desk",
    "version": "1.0.0",
    "description": "Refundable holds for agents renting real-world things. The agent holds $1.00 USDC with x402 upto on Solana, the desk checks whether the need is already handled, then keeps the hold as the rental or returns it.",
    "x-guidance": "Read GET /v1/terms first (free). Then call POST /v1/rent/charger or POST /v1/rent/hotspot through a payment-aware client. The first response is a 402 x402 upto offer for a hold of up to $1.00; the client pays and retries, and the desk settles $1.00 when the rental is kept or $0.01 when the need is already handled. One call is one hold, so do not repeat a call that returned 200."
  },
  "paths": {
    "/v1/terms": {
      "get": {
        "operationId": "getTerms",
        "summary": "Read desk terms, refund rules and device public key",
        "description": "Free, no payment. Returns the rules the desk applies when it settles a hold and the devicePublicKey that verifies signed readings.",
        "responses": {
          "200": {
            "description": "Desk terms.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "service": {
                      "type": "string"
                    },
                    "summary": {
                      "type": "string"
                    },
                    "why_hold": {
                      "type": "string",
                      "description": "Why an agent holds $1.00 and what happens to it."
                    },
                    "network": {
                      "type": "string",
                      "description": "localnet in the Pay.sh sandbox, mainnet when live."
                    },
                    "scheme": {
                      "type": "string"
                    },
                    "devicePublicKey": {
                      "type": "string",
                      "description": "Hex ed25519 public key that verifies signed readings."
                    },
                    "items": {
                      "type": "object",
                      "description": "Per item hold_usd, check_fee_usd, check question and settlement rules."
                    },
                    "endpoints": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/rent/charger": {
      "post": {
        "operationId": "rentCharger",
        "summary": "Start a charger rental with a refundable $1 hold",
        "description": "Holds up to $1.00, then checks whether the device is already drawing AC power. Already on power: $0.01 check fee, $0.99 returned. On battery and power arrives within wait_seconds: $1.00 kept. Power never arrives: $0.01, $0.99 returned.",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "wait_seconds": {
                    "type": "number",
                    "minimum": 0,
                    "default": 30,
                    "description": "Seconds to wait for power to arrive when the device starts on battery. Keep it well under the 300 second payment timeout."
                  }
                }
              },
              "example": {
                "wait_seconds": 30
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Hold settled: kept as the rental or refunded, with a signed device reading.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "hold_id",
                    "item",
                    "outcome",
                    "decision",
                    "charged_usd",
                    "returned_usd",
                    "reason",
                    "signed_reading",
                    "network"
                  ],
                  "properties": {
                    "hold_id": {
                      "type": "string",
                      "description": "Short id of this hold."
                    },
                    "item": {
                      "type": "string",
                      "enum": [
                        "charger"
                      ]
                    },
                    "outcome": {
                      "type": "string",
                      "enum": [
                        "already_handled",
                        "delivered",
                        "not_delivered",
                        "check_failed"
                      ]
                    },
                    "decision": {
                      "type": "string",
                      "enum": [
                        "kept",
                        "refunded",
                        "settle_failed"
                      ],
                      "description": "kept or refunded once settled; settle_failed when the settlement transaction failed, in which case no money moved and settle_error says why."
                    },
                    "charged_usd": {
                      "type": "string",
                      "description": "USD settled to the desk, 1.00 or 0.01."
                    },
                    "returned_usd": {
                      "type": "string",
                      "description": "USD returned to the payer, 0.00 or 0.99."
                    },
                    "reason": {
                      "type": "string",
                      "description": "The refund rule that applied, in plain English."
                    },
                    "signed_reading": {
                      "type": "object",
                      "description": "Device reading signed with the desk device ed25519 key. A failed check returns an unsigned reading with outcome check_failed.",
                      "properties": {
                        "holdId": {
                          "type": "string"
                        },
                        "item": {
                          "type": "string"
                        },
                        "outcome": {
                          "type": "string"
                        },
                        "detail": {
                          "type": "string",
                          "description": "Short machine reason, for example device_on_AC or charging_verified."
                        },
                        "raw": {
                          "type": "string",
                          "description": "The raw device observation the decision was based on."
                        },
                        "ts": {
                          "type": "integer",
                          "description": "Unix time in milliseconds."
                        },
                        "signature": {
                          "type": "string",
                          "description": "Hex ed25519 signature over the reading fields."
                        },
                        "devicePublicKey": {
                          "type": "string",
                          "description": "Hex ed25519 public key, same value as GET /v1/terms devicePublicKey."
                        }
                      }
                    },
                    "settlement_tx": {
                      "type": "string",
                      "description": "Solana settlement transaction signature, present when available."
                    },
                    "network": {
                      "type": "string",
                      "description": "Network the hold settled on: localnet in the Pay.sh sandbox, mainnet when live."
                    },
                    "settle_error": {
                      "type": "string",
                      "description": "Present only when decision is settle_failed."
                    }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required. x402 upto offer for a hold of up to 1 USDC (1000000 base units)."
          }
        },
        "x-payment-info": {
          "offers": [
            {
              "amount": "1000000",
              "currency": "USDC",
              "description": "up to 1 USDC",
              "intent": "charge",
              "method": "x402",
              "network": "solana:5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp",
              "scheme": "upto"
            }
          ]
        }
      }
    },
    "/v1/rent/hotspot": {
      "post": {
        "operationId": "rentHotspot",
        "summary": "Start a hotspot rental with a refundable $1 hold",
        "description": "Holds up to $1.00, then checks whether the device is already on the venue network. Already on it: $0.01 check fee, $0.99 returned. Off the venue network: hotspot rental kept, $1.00 charged.",
        "responses": {
          "200": {
            "description": "Hold settled: kept as the rental or refunded, with a signed device reading.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "hold_id",
                    "item",
                    "outcome",
                    "decision",
                    "charged_usd",
                    "returned_usd",
                    "reason",
                    "signed_reading",
                    "network"
                  ],
                  "properties": {
                    "hold_id": {
                      "type": "string",
                      "description": "Short id of this hold."
                    },
                    "item": {
                      "type": "string",
                      "enum": [
                        "hotspot"
                      ]
                    },
                    "outcome": {
                      "type": "string",
                      "enum": [
                        "already_handled",
                        "delivered",
                        "check_failed"
                      ]
                    },
                    "decision": {
                      "type": "string",
                      "enum": [
                        "kept",
                        "refunded",
                        "settle_failed"
                      ],
                      "description": "kept or refunded once settled; settle_failed when the settlement transaction failed, in which case no money moved and settle_error says why."
                    },
                    "charged_usd": {
                      "type": "string",
                      "description": "USD settled to the desk, 1.00 or 0.01."
                    },
                    "returned_usd": {
                      "type": "string",
                      "description": "USD returned to the payer, 0.00 or 0.99."
                    },
                    "reason": {
                      "type": "string",
                      "description": "The refund rule that applied, in plain English."
                    },
                    "signed_reading": {
                      "type": "object",
                      "description": "Device reading signed with the desk device ed25519 key. A failed check returns an unsigned reading with outcome check_failed.",
                      "properties": {
                        "holdId": {
                          "type": "string"
                        },
                        "item": {
                          "type": "string"
                        },
                        "outcome": {
                          "type": "string"
                        },
                        "detail": {
                          "type": "string",
                          "description": "Short machine reason, for example device_on_AC or charging_verified."
                        },
                        "raw": {
                          "type": "string",
                          "description": "The raw device observation the decision was based on."
                        },
                        "ts": {
                          "type": "integer",
                          "description": "Unix time in milliseconds."
                        },
                        "signature": {
                          "type": "string",
                          "description": "Hex ed25519 signature over the reading fields."
                        },
                        "devicePublicKey": {
                          "type": "string",
                          "description": "Hex ed25519 public key, same value as GET /v1/terms devicePublicKey."
                        }
                      }
                    },
                    "settlement_tx": {
                      "type": "string",
                      "description": "Solana settlement transaction signature, present when available."
                    },
                    "network": {
                      "type": "string",
                      "description": "Network the hold settled on: localnet in the Pay.sh sandbox, mainnet when live."
                    },
                    "settle_error": {
                      "type": "string",
                      "description": "Present only when decision is settle_failed."
                    }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required. x402 upto offer for a hold of up to 1 USDC (1000000 base units)."
          }
        },
        "x-payment-info": {
          "offers": [
            {
              "amount": "1000000",
              "currency": "USDC",
              "description": "up to 1 USDC",
              "intent": "charge",
              "method": "x402",
              "network": "solana:5eykt4UsFv8P8NJdTREpY1vzqKqZKvdp",
              "scheme": "upto"
            }
          ]
        }
      }
    }
  }
}